Add new input argument to Model __init__
Add `refresh_seconds` as a new keyword argument to Model.__init__

The `progress_bar` attribute of `Model` is created from the
`progress_bar_factory` that takes an argument `refresh_seconds`. The
value of this argument is hard-coded at 0.5. This creates a frequent
updating of progress, which creates massive amounts of outputs for long
running simulations. It would be preferrable to override this value when
creating a new model.